Refactor location of BwTransaction
