+++ packages/bierzo-wallet/src/logic/transactions/types/BwTransaction.ts
